Adjusting to Google-style III
