+++ src/main/java/mops/portfolios/PortfoliosApplication.java
@@ -4,10 +4,10 @@ import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import mops.portfolios.domain.portfolio.Portfolio;
 import mops.portfolios.domain.portfolio.PortfolioRepository;
-import mops.portfolios.domain.userGroup.User;
-import mops.portfolios.domain.userGroup.Group;
-import mops.portfolios.domain.userGroup.UserGroup;
-import mops.portfolios.domain.userGroup.UserGroupRepository;
+import mops.portfolios.domain.usergroup.User;
+import mops.portfolios.domain.usergroup.Group;
+import mops.portfolios.domain.usergroup.UserGroup;
+import mops.portfolios.domain.usergroup.UserGroupRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.CommandLineRunner;
