--- src/main/java/mops/portfolios/domain/userGroup/UserGroupRepository.java
@@ -1,12 +1,0 @@
-package mops.portfolios.domain.userGroup;
-
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.List;
-
-public interface UserGroupRepository extends CrudRepository<UserGroup, Long>
-{
-    List<UserGroup> findAllByUserId(String userId);
-
-    List<UserGroup> findAllByGroupId(Long groupId);
-}
