+++ src/main/java/mops/portfolios/domain/usergroup/Group.java
@@ -1,0 +1,24 @@
+package mops.portfolios.domain.usergroup;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class Group {
+  private Long id;
+  private String title;
+
+  public Group(Long id, String title) {
+    this.id = id;
+    this.title = title;
+  }
+
+  public String getTitle() {
+    return title;
+  }
+
+  public Long getId() {
+    return id;
+  }
+}
