+++ src/main/java/mops/portfolios/domain/usergroup/User.java
@@ -1,7 +1,6 @@
-package mops.portfolios.domain.userGroup;
+package mops.portfolios.domain.usergroup;
 
 import java.util.Set;
-
 import lombok.Data;
 
 @Data
@@ -11,6 +10,14 @@ public class User {
   private final String image;
   private final Set<String> roles;
   private final String id;
+
+  /** All args constructor with the following variables.
+   * @param name - User name
+   * @param email - User email
+   * @param image - User profile picture
+   * @param roles - Student or Orga
+   * @param id - User id
+   */
 
   public User(String name, String email, String image, Set<String> roles, String id) {
     this.name = name;
